Make the repo public
Could be considered as a mistake. Initially I wanted to protect the translation projects by avoiding having multiple translations in the same language at the same time. Since it's happening anyway (with the Italian one) and that the other ones are pretty much too advanced at the moment, there's no reason it should happen now.